refactor: letter service 패키지 분리
